# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Документация API к приложению достопримечательностей",
        "description": "Авторизация происходит через OAuth2, описание и ссылки на методы надоятся в Authorize",
        "contact": {
            "email": "sergei.klimovich.com"
        },
        "version": "0.1"
    },
    "servers": [
        {
            "url": "https://travel.madskill.ru",
            "description": "API сервер"
        }
    ],
    "paths": {
...
        "/api/attraction": {
            "get": {
                "tags": [
                    "Достопримечательности"
                ],
                "summary": "Одна Достопримечательность",
                "description": "Одна Достопримечательность",
                "operationId": "attraction",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SingleAttraction"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Текст ошибки"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/profile": {
            "get": {
                "tags": [
                    "Профиль пользователя"
                ],
                "summary": "Профиль",
                "description": "Профиль",
                "operationId": "profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ваш профиль не заполнен"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Профиль пользователя"
                ],
                "summary": "Изменение Профиля",
                "description": "Изменение Профиля",
                "operationId": "postprofile",
                "parameters": [
                    {
                        "name": "city",
                        "in": "query",
                        "description": "город",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Имя",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "email",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ваш профиль успешно заполнен"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ваш профиль не заполнен"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Профиль пользователя"
                ],
                "description": "Загрузка фото",
                "operationId": "76ee035e127a707b7598b1ad77005809",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "description": "file to upload",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Вы не авторизованы"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/visit": {
            "get": {
                "tags": [
                    "Профиль пользователя"
                ],
                "summary": "Посещенные достопримечательности",
                "description": "Посещенные достопримечательности",
                "operationId": "visit",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Visit"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ваш профиль не заполнен"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Профиль пользователя"
                ],
                "summary": "Изменение посещений",
                "description": "Изменение посещений",
                "operationId": "postvisit",
                "parameters": [
                    {
                        "name": "attractionID",
                        "in": "query",
                        "description": "id достопримечательности",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Посещение добавлено"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Достопримечательность не найдена"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "423": {
                        "description": "Ошибка",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "id не может быть пустым"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "424": {
                        "description": "Ошибка",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Вы не авторизованы"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Photo": {
                "title": "Photo",
                "description": "Модель фотографий",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "attraction_id": {
                        "title": "attraction_id",
                        "description": "attraction_id",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "image": {
                        "title": "image",
                        "description": "Название",
                        "type": "string",
                        "format": "string",
                        "example": "пример"
                    }
                },
                "type": "object"
            },
...
            "SingleAttraction": {
                "title": "SingleAttraction",
                "description": "Модель одной достопримечательности",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "name",
                        "description": "Название",
                        "type": "string",
                        "format": "string",
                        "example": "пример"
                    },
                    "description": {
                        "title": "description",
                        "description": "Описание",
                        "type": "string",
                        "format": "string",
                        "example": "пример"
                    },
                    "city": {
                        "title": "city",
                        "description": "Город",
                        "type": "string",
                        "format": "string",
                        "example": "пример"
                    },
                    "photo": {
                        "title": "photo",
                        "description": "photo",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Photo"
                        }
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "description": "Модель пользователя",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "name",
                        "description": "name",
                        "type": "string",
                        "format": "string",
                        "example": 1
                    },
                    "email_verified_at": {
                        "title": "email_verified_at",
                        "description": "email_verified_at",
                        "type": "string",
                        "format": "string",
                        "example": 1
                    },
                    "created_at": {
                        "title": "created_at",
                        "description": "created_at",
                        "type": "string",
                        "format": "string",
                        "example": 1
                    },
                    "updated_at": {
                        "title": "updated_at",
                        "description": "updated_at",
                        "type": "string",
                        "format": "string",
                        "example": 1
                    },
                    "profile": {
                        "$ref": "#/components/schemas/Profile"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Внимание OAuth2 scopes none, grant type code",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "authorizationCode": {
                        "authorizationUrl": "https://travel.madskill.ru/oauth/authorize",
                        "tokenUrl": "https://travel.madskill.ru/oauth/token",
                        "refreshUrl": "https://travel.madskill.ru/token/refresh",
                        "redirectUrl": "https://travel.madskill.ru/auth/callback",
                        "scopes": []
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}